% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apisrs}
\alias{apisrs}
\title{Student performance in California schools}
\format{A data frame with 39 variables including: \code{sytpe}, \code{cname},
\code{api00}, \code{pcttest}, \code{awards},
\code{sch.wide}
\describe{
\item{stype}{Elementary/Middle/High School}
\item{cname}{County name}
\item{api00}{Academic Performance Index in 2000}
\item{pcttest}{percentage of students tested}
\item{awards}{Eligible for awards program}
\item{sch.wide}{Met school-wide growth target}
}}
\usage{
apisrs
}
\description{
Source: http://www.ats.ucla.edu/stat/stata/Library/svy_survey.htm
}
\keyword{datasets}
