% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_tables.R
\name{basic_surv_prop_table}
\alias{basic_surv_prop_table}
\title{A workhourse function to construct survey crosstabs}
\usage{
basic_surv_prop_table(s, ..., binary = F)
}
\arguments{
\item{s}{A tbl_svy object created using the srvyr package.}

\item{...}{Names columns to be used in as grouping variables followed by
the column to be used as the outcome variable}

\item{binary}{A logical indicating whether the outcome variable is
binary and hence only one column should be displayed}
}
\value{
A long format tibble. Each of the initial character columns is named
for a grouping variable and list combination of grouping categories for that row
The later numeric columns are named for the outcome categories and list the
percentages falling into each. Rows should sum to 100, unless outcome is binary
in which case only the column for "TRUE" is displayed to avoid redudancy.
}
\description{
A workhourse function to construct survey crosstabs
}
\examples{
data(ca_school_tbl_svy)
ca_school_tbl_svy \%>\% basic_surv_prop_table("sch.wide.imp.goal", "comparative.imp.goal", "school.level")
}
