% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_school_svy}
\alias{ca_school_svy}
\title{Student performance in California schools (srvyr object).}
\format{A simple probability survey with 200 observations of 11 variables:
\describe{
\item{cds}{Unique identifier}
\item{fpc}{finite population correct}
\item{stype}{Elementary/Middle/High School}
\item{cname}{County name}
\item{pcttest}{percentage of students tested}
\item{ell}{Percent English language learners}
\item{meals}{Percent of students on free and reduced price lunch}
\item{poverty}{Federally-defined poverty level based meals variable}
\item{winner}{Eligible for awards program}
\item{sch.wide}{Met school-wide growth target}
\item{comp.imp}{Met Comparable Improvement target}
}}
\source{
\url{http://www.ats.ucla.edu/stat/stata/Library/svy_survey.htm}
}
\usage{
ca_school_svy
}
\description{
Built from the \code{ca_school_data} data frame, which in turn derives from \code{apistrat}
in the \code{api} data in the survey package.
}
\keyword{datasets}
