% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map_selected_projects.R
\name{draw_map_selected_projects}
\alias{draw_map_selected_projects}
\title{Draw map of selected projects}
\usage{
draw_map_selected_projects(projects_data_sf, zoom_level = 8, language)
}
\arguments{
\item{projects_data_sf}{A sf data.frame containing coordinates of project
main organisation and project `short_title`.}

\item{zoom_level}{An integer. The zoom level for the map. Defines
default zoom level.}

\item{language}{A character string. App language "fr", "de"...}
}
\value{
A leaflet object.
}
\description{
Draw map of selected projects
}
