% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratorium.R
\docType{package}
\name{exploratorium}
\alias{exploratorium}
\alias{exploratorium-package}
\title{exploratorium}
\description{
The exploratorium package is designed for exploratory data analysis
of weighted survey data
}
