% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_tables.R
\name{intermediate_surv_prop_table}
\alias{intermediate_surv_prop_table}
\title{An intermediate helper function for surv_prop_table}
\usage{
intermediate_surv_prop_table(s, outcome_vars, grouping_vars,
  new_outcome_name = "outcome", binary = F)
}
\arguments{
\item{s}{A tbl_svy object created using the srvyr package.}

\item{outcome_vars}{A character vector naming the column(s) in \code{s} to be used
as outcomes variable(s). These must be binary or categorial columns in s (convert
beforehand if needed).}

\item{grouping_vars}{A character vector naming the column(s) in \code{s} to be used
as grouping variable(s). These must be binary or categorial columns in s
(convert beforehand if needed).}

\item{new_outcome_name}{Character with intended name of the column
to hold the names of the outcome variables. Ignored if only 1 outcome.}

\item{binary}{Logical indicating whether the outcome(s) are binary
in which case only one column of outcome values will be displayed.}
}
\value{
A long format tibble. Each of the initial character columns is named
for a grouping variable and list combination of grouping categories for that row
The later numeric columns are named for the outcome categories and list the
percentages falling into each. Rows should sum to 100, unless outcome is binary
in which case only the column for "TRUE" is displayed to avoid redudancy.
}
\description{
An intermediate helper function for surv_prop_table
}
\examples{
data(ca_school_tbl_svy)
# Multiple grouping variables
ca_school_tbl_svy \%>\%
  intermediate_surv_prop_table(outcome_vars = "school.level",
                 grouping_vars = c("sch.wide.imp.goal", "comparative.imp.goal"))

# Multiple outcomes
ca_school_tbl_svy \%>\%
 intermediate_surv_prop_table(outcome_vars = c("sch.wide.imp.goal", "comparative.imp.goal"),
                grouping_vars = "school.level",
                new_outcome_name = "met_standards")

# Binary outcome
ca_school_tbl_svy \%>\%
 intermediate_surv_prop_table(outcome_vars = "eligible.for.award",
                grouping_vars = "school.level",
                binary = T)
}
