% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-moe.R
\name{moe}
\alias{moe}
\title{Margin of error}
\usage{
moe(est, n = 400, confidence = 95)
}
\arguments{
\item{est}{The point estimate as a percent.}

\item{n}{The sample size.}

\item{confidence}{The confidence level as a percent}
}
\value{
The margin of error in percentage points.
}
\description{
Calcuates the margin of error.
}
\details{
Under the usual frequentist assumptions, the underlying percentage
of the quantity of interested in the population ought to be within
± the returned value of the original estimate
}
\examples{
moe(50, 750, 95)
}
