% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pct.R
\name{pct}
\alias{pct}
\title{Convert proportions to percents}
\usage{
pct(x, decimals = 1)
}
\arguments{
\item{x}{numeric vector or data.frame}

\item{decimals}{integer indicating how many decimal places to round to}
}
\value{
object of same class as x
}
\description{
This function converts decimals to precents and rounds them
}
\examples{
vec <- c(0.4, 0.0265, 0.3234)
pct(vec, decimals = 1)
}
