% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_budget_barchart.R
\name{plot_budget_barchart}
\alias{plot_budget_barchart}
\title{Plot the distribution of the budgets with a barchart}
\usage{
plot_budget_barchart(
  id,
  data_repart,
  x_axis_labels = "true",
  axis_max = 100,
  axis_interval = 20,
  font_size_labels_col = "16px",
  series_name = "",
  prefixer = "\%",
  session
)
}
\arguments{
\item{id}{Character. Id of the container in the DOM.}

\item{data_repart}{Tibble. Data to be plotted.}

\item{x_axis_labels}{Character. Should the labels on the x axis be displayed.}

\item{axis_max}{Integer. Maximum value for the axis.}

\item{axis_interval}{Integer. Interval for the axis.}

\item{font_size_labels_col}{Character. Font size for the labels next to the bars.}

\item{series_name}{Character. Series name.}

\item{prefixer}{Character. Prefixer.}

\item{session}{Session object.}
}
\value{
A Java Script plot
}
\description{
Plot the distribution of the budgets with a barchart
}
