% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contrib_budget_highcharter.R
\name{plot_contrib_budget_highcharter}
\alias{plot_contrib_budget_highcharter}
\title{Plot the distribution of the budgets}
\usage{
plot_contrib_budget_highcharter(
  id,
  data_repart,
  plot_options = list(axis_max = 100, axis_interval = 20, series_name = "", prefixer =
    "\%"),
  session
)
}
\arguments{
\item{id}{Character. Id of the container in the DOM.}

\item{data_repart}{Tibble. Data to be plotted.}

\item{plot_options}{List. List of options for the plot.}

\item{session}{Session object.}
}
\value{
A Java Script plot
}
\description{
Plot the distribution of the budgets
}
\examples{
toy_data_repart <- tibble::tibble(
  name = c("Item 1", "Item 2"),
  value = c(75, 25), 
  value_tooltip = c("75 000 CHF", "25 000 CHF")
)

# plot_contrib_budget_highcharter(
#   id = "id-in-the-dom-maybe-with-ns",
#   data_repart = toy_data_repart,
#   session = session
# )
}
