% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_projects_cards.R
\name{prepare_projects_cards}
\alias{prepare_projects_cards}
\title{Prepare all the project cards in DE and FR}
\usage{
prepare_projects_cards(
  data_projects_fr = read_projects_data(language = "fr"),
  data_projects_de = read_projects_data(language = "de"),
  pkg_dir = system.file(package = "observatoire")
)
}
\arguments{
\item{data_projects_fr}{Tibble. Projects data in FR.}

\item{data_projects_de}{Tibble. Projects data in DE.}

\item{pkg_dir}{Character. Path to the package.}
}
\value{
Nothing. Used for side effects. Create the project card.
}
\description{
Prepare all the project cards in DE and FR
}
\examples{
# Load the toy datasets
data("toy_projects_data_sf")

# Keep 2 projects
toy_data <- toy_projects_data_sf[1:2, ]

# Create a temp folder with data-projects-cards subfolder
my_temp_dir <- tempfile("test-create-project-card")
dir.create(my_temp_dir)
dir.create(file.path(my_temp_dir, "data-projects-cards"))
  
file.copy(
  from = system.file(
    "data-projects-cards", 
    "template_projects_cards.html", 
    package = "observatoire"
    ), 
  to = file.path(
    my_temp_dir, 
    "data-projects-cards", 
    "template_projects_cards.html"
    )
)
  
prepare_projects_cards(
  data_projects_fr = toy_data,
  data_projects_de = toy_data,
  pkg_dir = my_temp_dir
)

list.files(
  file.path(
    my_temp_dir, 
    "data-projects-cards"
  )
)

unlink(my_temp_dir, recursive = TRUE)
}
