% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_projects_data.R
\name{prepare_projects_data}
\alias{prepare_projects_data}
\title{Prepare the projects data in FR and DE to be included in the app as .rds files}
\usage{
prepare_projects_data(
  name_raw_file = "PGV.xlsx",
  pkg_dir = system.file(package = "observatoire")
)
}
\arguments{
\item{name_raw_file}{Character. Name of the raw data file.}

\item{pkg_dir}{Character. Path to the package (must contain a data-raw folder).}
}
\value{
Nothing to return, used for side effect. A message is displayed in the console.
}
\description{
Prepare the projects data in FR and DE to be included in the app as .rds files
}
\examples{
prepare_projects_data()
}
