% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_estimation.R
\name{simple_binary_tab}
\alias{simple_binary_tab}
\title{Create a table estimating the percentage of "yes" replies to a yes/no survey question
and the margin of error}
\usage{
simple_binary_tab(dat, binary_outcome, na.rm = T, vartype = NULL)
}
\arguments{
\item{dat}{A tbl_svy object created using the srvyr package}

\item{binary_outcome}{The name of an outcome variable in \code{dat}. No quotes needed.}

\item{na.rm}{A logical indicating whether to drop NAs from the outcome.}

\item{vartype}{A string indicting how to display the variance.
See \code{srvyr::survey_mean()}. "ci" is preferable.}
}
\value{
tibble with up to three numeric columns showing point estimate,
and uncertainty (e.g., variance or upper and lower bounds of the 95%
confidence interval)
}
\description{
Create a table estimating the percentage of "yes" replies to a yes/no survey question
and the margin of error
}
\examples{
data(api)

my_tbl_svy <- apisrs \%>\% as_survey_design(ids = 1, fpc = fpc)

my_tbl_svy \%>\% mutate(elementary = stype=="E") \%>\%
    simple_binary_tab(elementary, vartype = "ci")

rm(apiclus1, apiclus2, apipop, apisrs, apistrat, my_tbl_svy)
}
