% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_tables.R
\name{surv_mean_table}
\alias{surv_mean_table}
\title{Create crosstab of means and confidence intervals for numeric outcome(s)}
\usage{
surv_mean_table(s, outcome_vars, grouping_vars = NULL, na.rm = T,
  displayN = T, prop.to.pct = F, wrap = F)
}
\arguments{
\item{s}{A \code{tbl_svy} object created with the srvyr package.}

\item{outcome_vars}{A character vector of outcome variable names. The elements are the
existing names of numeric or logical columns to take the means of. If it is
a named vector, the names are what they variables should be renamed as.}

\item{grouping_vars}{Name of factor or character column to group by}

\item{na.rm}{A logical indicating whether to drop NAs from output table}

\item{displayN}{A logical indicating whether to append the number of
observations to each group's name. Ignored if there are zero or multiple
grouping variables.}

\item{prop.to.pct}{A logical indicating whether the outcome_vars are
proportions. If \code{TRUE}, convert estimates to percents and truncate
confidence intervals at 0 and 100}

\item{wrap}{A logical indicating to return html code that will print
the data to the Viewer window in Rstudio for copy and pasting into
other programs.}
}
\value{
A long formate tibble, or if wrap = T, a kableExtra table in html.
}
\description{
Takes in a \code{tbl_svy} built with the srvyr package, one or more grouping
}
\examples{
data(ca_school_tbl_svy)

# one outcome, no groups
surv_mean_table(ca_school_tbl_svy, "qual.teach.prop")

# multiple outcomes, no groups

outcome_vec <- c("eng.lang.learn.prop", "qual.teach.prop", "new.stu.prop")
surv_mean_table(ca_school_tbl_svy, outcome_vec)

# if you plan to rename these outcome variables, create a named vector
  # and let the function rename them to avoid repetition

named_outcome_vec <- c(`Fully qualified teachers` ="qual.teach.prop",
                      `New students`= "new.stu.prop",
                      `English language learners` = "eng.lang.learn.prop")

surv_mean_table(ca_school_tbl_svy, named_outcome_vec)

# prepare for copying or export using wrap (i.e., knitr::knit)
surv_mean_table(ca_school_tbl_svy, named_outcome_vec, wrap = T)

# one grouping varible
surv_mean_table(ca_school_tbl_svy, named_outcome_vec, "poverty")

# one grouping varible, N's not appended to names
surv_mean_table(ca_school_tbl_svy, named_outcome_vec, "poverty", displayN = F)

# multiple grouping variables (note that N's can't be appended)
surv_mean_table(ca_school_tbl_svy, named_outcome_vec, c("poverty", "size"))

# turn all columns into percentages and trim the CIs and 0 and 100
surv_mean_table(ca_school_tbl_svy, named_outcome_vec, c("poverty", "size"),
                 prop.to.pct =T)
rm(ca_school_tbl_svy)
}
