% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_tables.R
\name{surv_prop_table}
\alias{surv_prop_table}
\title{Create easy-to-export survey crosstabs}
\usage{
surv_prop_table(s, outcome_vars, grouping_vars = NULL,
  filter_vars = NULL, new_outcome_name = "outcome",
  new_group_name = "group", wrap = F)
}
\arguments{
\item{s}{A tbl_svy object created using the srvyr package.}

\item{outcome_vars}{A character vector naming the column(s) in \code{s} to be used
as outcomes variable(s). These must be binary or categorial columns in s (convert
beforehand if needed).}

\item{grouping_vars}{A character vector naming the column(s) in \code{s} to be used
as grouping variable(s). These must be binary or categorial columns in s
(convert beforehand if needed).}

\item{filter_vars}{character or character vector with names of
filter varibles. These must be binary columns in s of type logical or numeric
(convert beforehand if needed). Values equal to 1 (true) are retained.
They filters will be applied seperately, not in succession,
in order to allow one to create multiple overlapping groups.
If more elaborate filtering is required, use filter( ) beforehand.}

\item{new_outcome_name}{Character with intended name of the column
to hold the names of the outcome variables. Ignored if only 1 outcome.}

\item{new_group_name}{A character with intended name of the column to
hold the names of the filter variables being treated as groups.
Ignored if less than 2 filter_vars.}

\item{wrap}{A logical indicating to return html code that will print
the data to the Viewer window in Rstudio for copy and pasting into
other programs.}
}
\value{
A long formate tibble, or if wrap = T, a kableExtra table in html.
The intial columns are named for the grouping variables (including new groups
created from multiple outcomes and filters) and list all of a row's group
memberships. The later columns are named for the outcome categories and list the
percentages for each. Rows should sum to 100, unless outcome is binary
in which case only one column is displayed to avoid redudancy.
}
\description{
This function creates crosstabs for any number of grouping, filter,
and outcome variables.
}
\details{
This function outputs a wide format table with a column for each
grouping variable listing its categories and a column of values for each
response category of the outcome(s). Rows will sum to 100%
If outcomes are binary only one column of values is shown
If there are multiple outcomes, an outcome column is added to indicate
which outcome is being used for each row. If multiple filter variables,
a filter column is added to indicated which filter is used on each row
(this allows you to create groups that overlap)
}
\examples{
data(ca_school_tbl_svy)

# No groups
ca_school_tbl_svy \%>\%
  surv_prop_table("poverty",
  wrap = T)

ca_school_tbl_svy \%>\%
  surv_prop_table("poverty", "size",
  wrap = T)

# One grouping, binary outcome (only the "yes" column will be retained)
ca_school_tbl_svy \%>\%
  surv_prop_table("eligible.for.award", "school.level",
  wrap = T)

# Multiple groupings
ca_school_tbl_svy \%>\%
  surv_prop_table(outcome_vars = "poverty",
                 grouping_vars = c("size", "school.level"),
                 wrap = T)

# Multiple groupings and outcomes
# Note: these outcomes are yes/no variables, so only the "yes" column will be retained
ca_school_tbl_svy \%>\%
 surv_prop_table(outcome_vars = c("sch.wide.imp.goal", "comparative.imp.goal"),
                grouping_vars = c("size", "school.level"),
                new_outcome_name = "standards met",
                wrap = T)

# Overlapping groupting variables
ca_school_tbl_svy \%>\%
  mutate(sch.wide.imp.goal = sch.wide.imp.goal == "Yes",
         comparative.imp.goal = comparative.imp.goal == "Yes") \%>\%
 surv_prop_table(outcome_vars = "poverty",
           filter_vars = c("sch.wide.imp.goal","comparative.imp.goal"),
           new_group_name = "standards met",
           wrap = T)

# Delay wrapping in order to modify the output first
ca_school_tbl_svy \%>\%
  surv_prop_table(outcome_vars = c("sch.wide.imp.goal", "comparative.imp.goal"),
                grouping_vars = "school.level",
                new_outcome_name = "standards met",
                wrap = F) \%>\%
 mutate(`standards met` =
          recode(`standards met`,
                 sch.wide.imp.goal = "schoolwide target",
                 comparative.imp.goal = "comparative improvement target")) \%>\%
  rename(`school level` = school.level) \%>\%
  wrap()
rm(ca_school_tbl_svy)
}
