% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_estimation.R
\name{survestimates}
\alias{survestimates}
\title{Create crosstab of means and confidence intervals for numeric outcome(s)}
\usage{
survestimates(.s, .named_outcome_vec, .grp_var_1 = NULL, ...,
  .na.rm = T, .displayN = T, .proportions = F, .wrap = F)
}
\arguments{
\item{.named_outcome_vec}{A named character vector. The elements are the
existing names of numeric or logical columns to take the means of. The names
are what they should be renamed as.}

\item{.grp_var_1}{Name of factor or character column to group by}

\item{...}{Option additional factor or character columns to group by}

\item{.na.rm}{A logical indicating whether to drop NAs from output table}

\item{.displayN}{A logical indicating whether to append the number of
observations to each group's name. Ignored if there are zero or multiple
grouping variables.}

\item{.proportions}{A logical indicating whether the outcomes are
proportions. If \code{TRUE}, convert estimates to percents and truncate
confidence intervals at 0 and 100}

\item{.data}{A \code{tbl_svy} object created with the srvyr package.}

\item{wrap}{A logical indicating to return html code that will print
the data to the Viewer window in Rstudio for copy and pasting into
other programs.}
}
\value{
A long formate tibble, or if wrap = T, a kableExtra table in html.
}
\description{
Takes in a \code{tbl_svy} built with the srvyr package, one or more grouping
}
\examples{
data(ca_school_svy)

named_outcome_vec <- c(`percent tested` ="pcttest",
                       `free and reduced price lunch`= "meals",
                       `English language learners` = "ell")
# no grouping variables
survestimates(ca_school_svy, named_outcome_vec)

# one grouping varible
survestimates(ca_school_svy, named_outcome_vec, comp.imp, stype, winner)

# multiple grouping variables
survestimates(ca_school_svy, named_outcome_vec, comp.imp)
}
