% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_toy_cantons_sf.R
\docType{data}
\name{toy_cantons_sf}
\alias{toy_cantons_sf}
\title{toy_cantons_sf}
\format{
A data frame with 26 rows and 12 variables:
\describe{
  \item{ GID_1 }{  character }
  \item{ GID_0 }{  character }
  \item{ COUNTRY }{  character }
  \item{ NAME_1 }{  character }
  \item{ VARNAME_1 }{  character }
  \item{ NL_NAME_1 }{  character }
  \item{ TYPE_1 }{  character }
  \item{ ENGTYPE_1 }{  character }
  \item{ CC_1 }{  character }
  \item{ HASC_1 }{  character }
  \item{ ISO_1 }{  character }
  \item{ geometry }{  sfc_MULTIPOLYGON,sfc }
}
}
\source{
Source
}
\usage{
toy_cantons_sf
}
\description{
Description.
}
\keyword{datasets}
