% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_toy_dic_cantons.R
\docType{data}
\name{toy_dic_cantons}
\alias{toy_dic_cantons}
\title{toy_dic_cantons}
\format{
A data frame with 27 rows and 3 variables:
\describe{
  \item{ id }{  character }
  \item{ fr }{  character }
  \item{ de }{  character }
}
}
\source{
Source
}
\usage{
toy_dic_cantons
}
\description{
Description.
}
\keyword{datasets}
