% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_toy_dic_values.R
\docType{data}
\name{toy_dic_values}
\alias{toy_dic_values}
\title{toy_dic_values}
\format{
A data frame with 63 rows and 3 variables:
\describe{
  \item{ value }{  character }
  \item{ de }{  character }
  \item{ fr }{  character }
}
}
\source{
Source
}
\usage{
toy_dic_values
}
\description{
Description.
}
\keyword{datasets}
