% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_toy_dic_variables.R
\docType{data}
\name{toy_dic_variables}
\alias{toy_dic_variables}
\title{toy_dic_variables}
\format{
A data frame with 42 rows and 4 variables:
\describe{
  \item{ name_col_in_raw_data }{  character }
  \item{ de }{  character }
  \item{ fr }{  character }
  \item{ name_variable }{  character }
}
}
\source{
Source
}
\usage{
toy_dic_variables
}
\description{
Description.
}
\keyword{datasets}
