% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_toy_projects_data_sf.R
\docType{data}
\name{toy_projects_data_sf}
\alias{toy_projects_data_sf}
\title{toy_projects_data_sf}
\format{
A data frame with 67 rows and 51 variables:
\describe{
  \item{ short_title }{  character }
  \item{ language }{  character }
  \item{ funding_round }{  character }
  \item{ funding_area }{  character }
  \item{ project_start }{  POSIXct,POSIXt }
  \item{ project_end }{  POSIXct,POSIXt }
  \item{ description }{  character }
  \item{ geo_range }{  character }
  \item{ main_resp_orga }{  character }
  \item{ zip_code_main_resp_orga }{  character }
  \item{ city_code_main_resp_orga }{  character }
  \item{ orga_type }{  character }
  \item{ project_support_gfch }{  character }
  \item{ budget_gfch }{  numeric }
  \item{ budget_orga }{  numeric }
  \item{ budget_third_party }{  numeric }
  \item{ total_budget }{  numeric }
  \item{ pi_1_interfaces }{  character }
  \item{ pi_1_health_pathways }{  character }
  \item{ pi_1_self_gestion }{  character }
  \item{ pi_2_training }{  character }
  \item{ pi_2_new_tech }{  character }
  \item{ pi_2_econom }{  character }
  \item{ target_groups }{  character }
  \item{ multipliers }{  character }
  \item{ settings }{  character }
  \item{ desc_fr }{  character }
  \item{ desc_de }{  character }
  \item{ title_de }{  character }
  \item{ title_fr }{  character }
  \item{ topic_addictions }{  logical }
  \item{ topic_cancers }{  logical }
  \item{ topic_cardio_diseases }{  logical }
  \item{ topic_diabetes }{  logical }
  \item{ topic_mental_diseases }{  logical }
  \item{ topic_musculoskeletal_diseases }{  logical }
  \item{ topic_other }{  logical }
  \item{ topic_other_ncd }{  logical }
  \item{ topic_respiratory_diseases }{  logical }
  \item{ latitude }{  numeric }
  \item{ longitude }{  numeric }
  \item{ geometry }{  sfc_POINT,sfc }
  \item{ id_canton }{  character }
  \item{ geo_range_id }{  character }
  \item{ nb_cantons_influenced }{  numeric }
  \item{ prop_budget_gfch }{  numeric }
  \item{ prop_budget_orga }{  numeric }
  \item{ prop_budget_third_party }{  numeric }
  \item{ status }{  character }
  \item{ topic }{  character }
  \item{ risk_factors }{  character }
}
}
\source{
Source
}
\usage{
toy_projects_data_sf
}
\description{
Description.
}
\keyword{datasets}
