% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap}
\alias{wrap}
\title{Create an neatly html formatted table}
\usage{
wrap(t)
}
\arguments{
\item{t}{A data.frame or similar rectangular object to format.}
}
\value{
html code for a sleek and simple table
}
\description{
Takes a rectangular data structure and outs and table that prints
to the Viewer screen in Rstudio for each coping to other programs
like Excel, GoogleSheets, or Datawrapper
}
\examples{
wrap(cars)
}
