% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_school_data}
\alias{ca_school_data}
\title{Student performance in California schools.}
\format{A data frame with 200 observations of 14 variables:
\describe{
\item{id}{Unique identifier}
\item{simple.weight}{Sampling Weight}
\item{year.round.sch}{Whether school runs year-round}
\item{eligible.for.award}{Whether the schools test score improvements from
1999 to 2000 rendered it eligible for an award}
\item{poverty}{Federally-defined poverty level based on proportion
of students on free and reduced-price meals}
\item{size}{Total enrollment is over 1000 (large), over 500 (medium),
or under 500 (small)}
\item{school.level}{Elementary/Middle/High School}
\item{sch.wide.imp.goal}{Met target for schoolwide growth}
\item{comparative.imp.goal}{Met target for comparative improvement}
\item{test.score.1999}{Average student test score in 1999}
\item{test.score.2000}{Average student test score in 2000}
\item{new.stu.prop}{Proportion of students who are new that year}
\item{eng.lang.learn.prop}{Proportion of English-language learners}
\item{qual.teach.prop}{Proportion of teachers who are fully qualified}
}}
\source{
\url{http://www.ats.ucla.edu/stat/stata/Library/svy_survey.htm}
}
\usage{
ca_school_data
}
\description{
Derived from \code{apistrat} data frame in the \code{api} data in the survey package.
}
\keyword{datasets}
