% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_tables.R
\name{cross_tab_plot}
\alias{cross_tab_plot}
\title{Plot multiple outcomes and confidence intervals efficiently}
\usage{
cross_tab_plot(tab, y_var = outcome, color_var = NULL,
  shape_var = NULL, dodge = 0.5, order_by_est = T, title = NA,
  subtitle = NA, dates = NA, caption = "default",
  x_axis_lab = element_blank(), y_axis_lab = element_blank(),
  x_scale = "default", color_scheme = "Dark2", x_min = 0,
  x_max = NA_real_)
}
\arguments{
\item{tab}{A table with columns named est, upp, and low
(for point estimate, upper CI bound, and lower CI bound)}

\item{y_var}{Name of outcome varible}

\item{color_var}{Name of grouping variable to differtiate by color (if any)}

\item{shape_var}{Name of grouping variable to differtiate by shape (if any)}

\item{dodge}{Numeric indicating the distance between estimates}

\item{order_by_est}{Logical determining whether to reorder outcomes accoring to
which was had the higher mean}

\item{title}{A string supplying the title. Often good to standarized half of
it at the top of the section or document and then customize it with paste()}

\item{subtitle}{A string supplying the subtitle which is typically the question
stem of survey questions that are being presented as outcomes}

\item{dates}{A string supplying the dates of the survey to go in the  caption.
Overridden by \code{caption}.}

\item{caption}{String. Useful if you want one caption for entire section or script.}

\item{x_axis_lab}{String w/ desired name of x-axis.}

\item{y_axis_lab}{String w/ desired name of y-axis.}

\item{x_scale}{String defaulting to "default" in which case the x-axis will
go from 0 to \code{x_max}.}

\item{color_scheme}{String supplying the color scheme.}

\item{x_max}{Numberic indicating highest deisred value for x-axis}
}
\value{
A ggplot
}
\description{
Plot multiple outcomes and confidence intervals efficiently
}
\examples{
data(ca_school_tbl_svy)

named_outcome_vec <- c(`Fully qualified teachers` ="qual.teach.prop",
                       `New students`= "new.stu.prop",
                       `English language learners` = "eng.lang.learn.prop")

tab <- surv_mean_table(ca_school_tbl_svy, named_outcome_vec,
                       c("poverty", "year.round.sch"),
                       prop.to.pct = T, wrap = F)

cross_tab_plot(tab, color_var = poverty, shape_var = year.round.sch,
               x_max = 110)

rm(tab)
}
