% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_app_data.R
\name{prepare_app_data}
\alias{prepare_app_data}
\title{Prepare the projects data in FR and DE to be included in the app as .rds files}
\usage{
prepare_app_data(
  name_raw_file = "PGV.xlsx",
  pkg_dir = system.file(package = "exploratorium"),
  dic_variables = NULL,
  dic_cantons = NULL,
  cantons_sf = NULL
)
}
\arguments{
\item{name_raw_file}{Character. Name of the raw data file.}

\item{pkg_dir}{Character. Path to the package (must contain a data-raw folder).}

\item{dic_variables}{Tibble. Variables dictionaries. Mainly used for examples and unit testing purpose.}

\item{dic_cantons}{Tibble. Canton dictionary. Mainly used for examples and unit testing purpose.}

\item{cantons_sf}{Sf data. Cantons geometry. Mainly used for examples and unit testing purpose.}
}
\value{
A list with the projects data in FR and DE.
}
\description{
Prepare the projects data in FR and DE to be included in the app as .rds files
}
\examples{
# Load the toy datasets
data("toy_data_pgv")
data("toy_dic_variables")
data("toy_dic_cantons")
data("toy_cantons_sf")

# Create a temp folder with data-projects-raw and ata-projects subfolder
my_temp_dir <- tempfile("test-prepare-data")
dir.create(my_temp_dir)
dir.create(file.path(my_temp_dir, "data-projects-raw"))
dir.create(file.path(my_temp_dir, "data-projects"))

# Save the toy PGV file inside
writexl::write_xlsx(
  toy_data_pgv,
  file.path(
    my_temp_dir,
    "data-projects-raw",
    "toy_PGV.xlsx"
  )
)

# Prepare the data
prepare_app_data(
  name_raw_file = "toy_PGV.xlsx",
  pkg_dir = my_temp_dir,
  dic_variables = toy_dic_variables,
  dic_cantons = toy_dic_cantons,
  cantons_sf = toy_cantons_sf
)

# Delete the tempdir
unlink(my_temp_dir, recursive = TRUE)
}
